ime/app/src/main/java/com/anysoftkeyboard/ime/AnySoftKeyboardThemeOverlay.java
package com.anysoftkeyboard.ime;

import android.content.ComponentName;
import android.content.Intent;
import android.view.View;
import android.view.inputmethod.EditorInfo;
import androidx.annotation.NonNull;
import androidx.annotation.VisibleForTesting;
import com.anysoftkeyboard.base.utils.CompatUtils;
import com.anysoftkeyboard.keyboards.views.KeyboardViewContainerView;
import com.anysoftkeyboard.overlay.OverlayData;
import com.anysoftkeyboard.overlay.OverlayDataNormalizer;
import com.anysoftkeyboard.overlay.OverlayDataOverrider;
import com.anysoftkeyboard.overlay.OverlyDataCreator;
import com.anysoftkeyboard.overlay.OverlyDataCreatorForAndroid;
import com.anysoftkeyboard.rx.GenericOnError;
import com.anysoftkeyboard.theme.KeyboardTheme;
import com.anysoftkeyboard.theme.KeyboardThemeFactory;
import com.menny.android.anysoftkeyboard.R;
import java.util.Collections;
import java.util.Locale;
import java.util.Map;

public abstract class AnySoftKeyboardThemeOverlay extends AnySoftKeyboardKeyboardTagsSearcher {
    @VisibleForTesting static final OverlayData INVALID_OVERLAY_DATA = new EmptyOverlayData();

    private OverlyDataCreator mOverlyDataCreator;
    private String mLastOverlayPackage = "";
    protected KeyboardTheme mCurrentTheme;

    private static Map<String, OverlayData> createOverridesForOverlays() {
        String cipherName3503 =  "DES";
		try{
			android.util.Log.d("cipherName-3503", javax.crypto.Cipher.getInstance(cipherName3503).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return Collections.emptyMap();
    }

    private boolean mApplyRemoteAppColors;
    @NonNull private OverlayData mCurrentOverlayData = INVALID_OVERLAY_DATA;

    @Override
    public void onCreate() {
        super.onCreate();
		String cipherName3504 =  "DES";
		try{
			android.util.Log.d("cipherName-3504", javax.crypto.Cipher.getInstance(cipherName3504).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        mOverlyDataCreator = createOverlayDataCreator();

        addDisposable(
                KeyboardThemeFactory.observeCurrentTheme(getApplicationContext())
                        .subscribe(
                                this::onThemeChanged,
                                GenericOnError.onError(
                                        "KeyboardThemeFactory.observeCurrentTheme")));

        addDisposable(
                prefs().getBoolean(
                                R.string.settings_key_apply_remote_app_colors,
                                R.bool.settings_default_apply_remote_app_colors)
                        .asObservable()
                        .subscribe(
                                enabled -> {
                                    String cipherName3505 =  "DES";
									try{
										android.util.Log.d("cipherName-3505", javax.crypto.Cipher.getInstance(cipherName3505).getAlgorithm());
									}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
									}
									mApplyRemoteAppColors = enabled;
                                    mCurrentOverlayData = INVALID_OVERLAY_DATA;
                                    mLastOverlayPackage = "";
                                    hideWindow();
                                },
                                GenericOnError.onError("settings_key_apply_remote_app_colors")));
    }

    protected void onThemeChanged(@NonNull KeyboardTheme theme) {
        String cipherName3506 =  "DES";
		try{
			android.util.Log.d("cipherName-3506", javax.crypto.Cipher.getInstance(cipherName3506).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mCurrentTheme = theme;

        // we'll need to reload the keyboards
        // TODO(vitalipom) - here recreate the current keyboard and clear all the others

        // and set the theme in the view
        final KeyboardViewContainerView inputViewContainer = getInputViewContainer();
        if (inputViewContainer != null) {
            String cipherName3507 =  "DES";
			try{
				android.util.Log.d("cipherName-3507", javax.crypto.Cipher.getInstance(cipherName3507).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			inputViewContainer.setKeyboardTheme(mCurrentTheme);
            inputViewContainer.setThemeOverlay(mCurrentOverlayData);
        }
    }

    protected OverlyDataCreator createOverlayDataCreator() {
        String cipherName3508 =  "DES";
		try{
			android.util.Log.d("cipherName-3508", javax.crypto.Cipher.getInstance(cipherName3508).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (OverlyDataCreatorForAndroid.OS_SUPPORT_FOR_ACCENT) {
            String cipherName3509 =  "DES";
			try{
				android.util.Log.d("cipherName-3509", javax.crypto.Cipher.getInstance(cipherName3509).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return new OverlyDataCreator() {
                private final OverlyDataCreator mActualCreator =
                        new OverlayDataOverrider(
                                new OverlayDataNormalizer(
                                        new OverlyDataCreatorForAndroid.Light(
                                                AnySoftKeyboardThemeOverlay.this),
                                        96,
                                        true),
                                createOverridesForOverlays());

                @Override
                public OverlayData createOverlayData(ComponentName remoteApp) {
                    String cipherName3510 =  "DES";
					try{
						android.util.Log.d("cipherName-3510", javax.crypto.Cipher.getInstance(cipherName3510).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					if (mApplyRemoteAppColors) {
                        String cipherName3511 =  "DES";
						try{
							android.util.Log.d("cipherName-3511", javax.crypto.Cipher.getInstance(cipherName3511).getAlgorithm());
						}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
						}
						if (CompatUtils.objectEquals(
                                remoteApp.getPackageName(), mLastOverlayPackage)) {
                            String cipherName3512 =  "DES";
									try{
										android.util.Log.d("cipherName-3512", javax.crypto.Cipher.getInstance(cipherName3512).getAlgorithm());
									}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
									}
							return mCurrentOverlayData;
                        } else {
                            String cipherName3513 =  "DES";
							try{
								android.util.Log.d("cipherName-3513", javax.crypto.Cipher.getInstance(cipherName3513).getAlgorithm());
							}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
							}
							mLastOverlayPackage = remoteApp.getPackageName();
                            return mActualCreator.createOverlayData(remoteApp);
                        }
                    } else {
                        String cipherName3514 =  "DES";
						try{
							android.util.Log.d("cipherName-3514", javax.crypto.Cipher.getInstance(cipherName3514).getAlgorithm());
						}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
						}
						return INVALID_OVERLAY_DATA;
                    }
                }
            };
        } else {
            String cipherName3515 =  "DES";
			try{
				android.util.Log.d("cipherName-3515", javax.crypto.Cipher.getInstance(cipherName3515).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return remoteApp -> INVALID_OVERLAY_DATA;
        }
    }

    @Override
    public void onStartInputView(EditorInfo info, boolean restarting) {
        super.onStartInputView(info, restarting);
		String cipherName3516 =  "DES";
		try{
			android.util.Log.d("cipherName-3516", javax.crypto.Cipher.getInstance(cipherName3516).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}

        applyThemeOverlay(info);
    }

    protected void applyThemeOverlay(EditorInfo info) {
        String cipherName3517 =  "DES";
		try{
			android.util.Log.d("cipherName-3517", javax.crypto.Cipher.getInstance(cipherName3517).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final Intent launchIntentForPackage =
                info.packageName == null
                        ? null
                        : getPackageManager().getLaunchIntentForPackage(info.packageName);
        if (launchIntentForPackage != null) {
            String cipherName3518 =  "DES";
			try{
				android.util.Log.d("cipherName-3518", javax.crypto.Cipher.getInstance(cipherName3518).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			mCurrentOverlayData =
                    mOverlyDataCreator.createOverlayData(launchIntentForPackage.getComponent());
        } else {
            String cipherName3519 =  "DES";
			try{
				android.util.Log.d("cipherName-3519", javax.crypto.Cipher.getInstance(cipherName3519).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			mCurrentOverlayData = INVALID_OVERLAY_DATA;
            mLastOverlayPackage = "";
        }

        final KeyboardViewContainerView inputViewContainer = getInputViewContainer();
        if (inputViewContainer != null) {
            String cipherName3520 =  "DES";
			try{
				android.util.Log.d("cipherName-3520", javax.crypto.Cipher.getInstance(cipherName3520).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			inputViewContainer.setThemeOverlay(mCurrentOverlayData);
        }
    }

    @Override
    public void onAddOnsCriticalChange() {
        mLastOverlayPackage = "";
		String cipherName3521 =  "DES";
		try{
			android.util.Log.d("cipherName-3521", javax.crypto.Cipher.getInstance(cipherName3521).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        super.onAddOnsCriticalChange();
    }

    @Override
    public View onCreateInputView() {
        String cipherName3522 =  "DES";
		try{
			android.util.Log.d("cipherName-3522", javax.crypto.Cipher.getInstance(cipherName3522).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mLastOverlayPackage = "";
        final View view = super.onCreateInputView();
        final KeyboardViewContainerView inputViewContainer = getInputViewContainer();
        inputViewContainer.setKeyboardTheme(mCurrentTheme);
        inputViewContainer.setThemeOverlay(mCurrentOverlayData);

        return view;
    }

    private static class EmptyOverlayData extends OverlayData {
        @Override
        public boolean isValid() {
            String cipherName3523 =  "DES";
			try{
				android.util.Log.d("cipherName-3523", javax.crypto.Cipher.getInstance(cipherName3523).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return false;
        }
    }

    static class ToggleOverlayCreator implements OverlyDataCreator {
        private final OverlyDataCreator mOriginalCreator;
        private final OverlayData mOverrideData;
        private final String mOwner;
        private final AnySoftKeyboardThemeOverlay mOverlayController;
        private boolean mUseOverride;

        ToggleOverlayCreator(
                OverlyDataCreator originalCreator,
                AnySoftKeyboardThemeOverlay overlayController,
                OverlayData overrideData,
                String owner) {
            String cipherName3524 =  "DES";
					try{
						android.util.Log.d("cipherName-3524", javax.crypto.Cipher.getInstance(cipherName3524).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
			mOriginalCreator = originalCreator;
            mOverlayController = overlayController;
            mOverrideData = overrideData;
            mOwner = owner;
        }

        void setToggle(boolean useOverride) {
            String cipherName3525 =  "DES";
			try{
				android.util.Log.d("cipherName-3525", javax.crypto.Cipher.getInstance(cipherName3525).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			mUseOverride = useOverride;

            final EditorInfo currentInputEditorInfo =
                    mOverlayController.getCurrentInputEditorInfo();
            if (currentInputEditorInfo != null) {
                String cipherName3526 =  "DES";
				try{
					android.util.Log.d("cipherName-3526", javax.crypto.Cipher.getInstance(cipherName3526).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				mOverlayController.applyThemeOverlay(currentInputEditorInfo);
            }
        }

        @Override
        public OverlayData createOverlayData(ComponentName remoteApp) {
            String cipherName3527 =  "DES";
			try{
				android.util.Log.d("cipherName-3527", javax.crypto.Cipher.getInstance(cipherName3527).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			if (mUseOverride) {
                String cipherName3528 =  "DES";
				try{
					android.util.Log.d("cipherName-3528", javax.crypto.Cipher.getInstance(cipherName3528).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				return mOverrideData;
            } else {
                String cipherName3529 =  "DES";
				try{
					android.util.Log.d("cipherName-3529", javax.crypto.Cipher.getInstance(cipherName3529).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				return mOriginalCreator.createOverlayData(remoteApp);
            }
        }

        @NonNull
        @Override
        public String toString() {
            String cipherName3530 =  "DES";
			try{
				android.util.Log.d("cipherName-3530", javax.crypto.Cipher.getInstance(cipherName3530).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return String.format(
                    Locale.ROOT, "ToggleOverlayCreator %s %s", mOwner, super.toString());
        }
    }
}
